+++ docs/apidiffs/current_vs_latest/opentelemetry-sdk-extension-jaeger-remote-sampler.txt
@@ -1,2 +1,10 @@
 Comparing source compatibility of  against 
-No changes.
+***! MODIFIED CLASS: PUBLIC FINAL io.opentelemetry.sdk.extension.trace.jaeger.sampler.JaegerRemoteSampler  (not serializable)
+	===  CLASS FILE FORMAT VERSION: 52.0 <- 52.0
+	---! REMOVED METHOD: PUBLIC(-) void close()
+---! REMOVED CLASS: PUBLIC(-) io.opentelemetry.sdk.extension.trace.jaeger.sampler.JaegerRemoteSamplerProvider  (not serializable)
+	---  CLASS FILE FORMAT VERSION: n.a. <- 52.0
+	---! REMOVED SUPERCLASS: java.lang.Object
+	---! REMOVED CONSTRUCTOR: PUBLIC(-) JaegerRemoteSamplerProvider()
+	---! REMOVED METHOD: PUBLIC(-) io.opentelemetry.sdk.trace.samplers.Sampler createSampler(io.opentelemetry.sdk.autoconfigure.ConfigProperties)
+	---! REMOVED METHOD: PUBLIC(-) java.lang.String getName()
