--- exporters/otlp/logging/src/main/java/io/opentelemetry/exporter/oltp/logging/OltpGrpcLoggingExporterBuilder.java
@@ -1,9 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.exporter.oltp.logging;
-
-/** */
-public class OltpGrpcLoggingExporterBuilder {}
