+++ exporters/otlp/logs/gradle.properties
