+++ exporters/otlp/logs/src/main/java/io/opentelemetry/exporter/oltp/logging/OtlpGrpcLoggingExporterBuilder.java
@@ -1,0 +1,133 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.exporter.oltp.logging;
+
+import io.grpc.ManagedChannel;
+import io.grpc.ManagedChannelBuilder;
+import io.grpc.Metadata;
+import io.grpc.stub.MetadataUtils;
+import javax.annotation.Nullable;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.time.Duration;
+import java.util.concurrent.TimeUnit;
+
+import static io.grpc.Metadata.ASCII_STRING_MARSHALLER;
+import static io.opentelemetry.api.internal.Utils.checkArgument;
+import static java.util.Objects.requireNonNull;
+
+/** Builder utility for this exporter. */
+public class OtlpGrpcLoggingExporterBuilder {
+
+  private static final String DEFAULT_ENDPOINT_URL = "http://localhost:4317";
+  private static final URI DEFAULT_ENDPOINT = URI.create(DEFAULT_ENDPOINT_URL);
+  private static final long DEFAULT_TIMEOUT_SECS = 10;
+
+  private ManagedChannel channel;
+  private long timeoutNanos = TimeUnit.SECONDS.toNanos(DEFAULT_TIMEOUT_SECS);
+  private URI endpoint = DEFAULT_ENDPOINT;
+
+  @Nullable private Metadata metadata;
+
+  /**
+   * Sets the managed channel to use when communicating with the backend. Takes precedence over
+   * {@link #setEndpoint(String)} if both are called.
+   *
+   * @param channel the channel to use
+   * @return this builder's instance
+   */
+  public OtlpGrpcLoggingExporterBuilder setChannel(ManagedChannel channel) {
+    this.channel = channel;
+    return this;
+  }
+
+  /**
+   * Sets the maximum time to wait for the collector to process an exported batch of metrics. If
+   * unset, defaults to {@value DEFAULT_TIMEOUT_SECS}s.
+   */
+  public OtlpGrpcLoggingExporterBuilder setTimeout(long timeout, TimeUnit unit) {
+    requireNonNull(unit, "unit");
+    checkArgument(timeout >= 0, "timeout must be non-negative");
+    timeoutNanos = unit.toNanos(timeout);
+    return this;
+  }
+
+  /**
+   * Sets the maximum time to wait for the collector to process an exported batch of metrics. If
+   * unset, defaults to {@value DEFAULT_TIMEOUT_SECS}s.
+   */
+  public OtlpGrpcLoggingExporterBuilder setTimeout(Duration timeout) {
+    requireNonNull(timeout, "timeout");
+    return setTimeout(timeout.toNanos(), TimeUnit.NANOSECONDS);
+  }
+
+  /**
+   * Sets the OTLP endpoint to connect to. If unset, defaults to {@value DEFAULT_ENDPOINT_URL}. The
+   * endpoint must start with either http:// or https://.
+   */
+  public OtlpGrpcLoggingExporterBuilder setEndpoint(String endpoint) {
+    requireNonNull(endpoint, "endpoint");
+
+    URI uri;
+    try {
+      uri = new URI(endpoint);
+    } catch (URISyntaxException e) {
+      throw new IllegalArgumentException("Invalid endpoint, must be a URL: " + endpoint, e);
+    }
+
+    if (uri.getScheme() == null
+        || (!uri.getScheme().equals("http") && !uri.getScheme().equals("https"))) {
+      throw new IllegalArgumentException(
+          "Invalid endpoint, must start with http:// or https://: " + uri);
+    }
+
+    this.endpoint = uri;
+    return this;
+  }
+
+  /**
+   * Add header to request. Optional. Applicable only if {@link
+   * OtlpGrpcLoggingExporterBuilder#endpoint} is set to build channel.
+   *
+   * @param key header key
+   * @param value header value
+   * @return this builder's instance
+   */
+  public OtlpGrpcLoggingExporterBuilder addHeader(String key, String value) {
+    if (metadata == null) {
+      metadata = new Metadata();
+    }
+    metadata.put(Metadata.Key.of(key, ASCII_STRING_MARSHALLER), value);
+    return this;
+  }
+
+  /**
+   * Constructs a new instance of the exporter based on the builder's values.
+   *
+   * @return a new exporter's instance
+   */
+  public OtlpGrpcLogsExporter build() {
+    if (channel == null) {
+      final ManagedChannelBuilder<?> managedChannelBuilder =
+          ManagedChannelBuilder.forTarget(endpoint.getAuthority());
+
+      if (endpoint.getScheme().equals("https")) {
+        managedChannelBuilder.useTransportSecurity();
+      } else {
+        managedChannelBuilder.usePlaintext();
+      }
+
+      if (metadata != null) {
+        managedChannelBuilder.intercept(MetadataUtils.newAttachHeadersInterceptor(metadata));
+      }
+
+      channel = managedChannelBuilder.build();
+    }
+    return new OtlpGrpcLogsExporter(channel, timeoutNanos);
+  }
+
+  OtlpGrpcLoggingExporterBuilder() {}
+}
