--- sdk-extensions/logging/src/main/java/io/opentelemetry/sdk/logging/LogSinkSdkProviderBuilder.java
@@ -1,14 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.logging;
-
-public final class LogSinkSdkProviderBuilder {
-  LogSinkSdkProviderBuilder() {}
-
-  public LogSinkSdkProvider build() {
-    return new LogSinkSdkProvider();
-  }
-}
