--- sdk-extensions/logging/src/main/java/io/opentelemetry/sdk/logging/data/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** The data format to model logs for export. */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.logging.data;
-
-import javax.annotation.ParametersAreNonnullByDefault;
