--- sdk-extensions/logging/src/main/java/io/opentelemetry/sdk/logging/package-info.java
@@ -1,14 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/**
- * The OpenTelemetry SDK implementation of logging.
- *
- * @see io.opentelemetry.sdk.logging.LogSinkSdkProvider
- */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.logging;
-
-import javax.annotation.ParametersAreNonnullByDefault;
