--- sdk/logging/src/main/java/io/opentelemetry/sdk/logging/LogEmitterSdk.java
@@ -1,8 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.logging;
-
-public class LogEmitterSdk {}
