--- sdk/logging/src/main/java/io/opentelemetry/sdk/logging/data/LoggingData.java
@@ -1,9 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.logging.data;
-
-/** */
-public abstract class LoggingData {}
