--- sdk/logging/src/main/java/io/opentelemetry/sdk/logging/export/README.md
@@ -1,2 +1,0 @@
-# OpenTelemetry logging export framework
-
