--- sdk/logging/src/main/java/io/opentelemetry/sdk/logging/export/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** Utilities that allow to export logging to the OpenTelemetry exporters. */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.logging.export;
-
-import javax.annotation.ParametersAreNonnullByDefault;
