+++ sdk/logs/gradle.properties
