+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/LogEmitterSdk.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.sdk.logs;
+
+public class LogEmitterSdk {}
