+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/LogSinkSdkProviderBuilder.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.sdk.logs;
+
+public final class LogSinkSdkProviderBuilder {
+  LogSinkSdkProviderBuilder() {}
+
+  public LogSinkSdkProvider build() {
+    return new LogSinkSdkProvider();
+  }
+}
