+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/data/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** The data format to model logs for export. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.logs.data;
+
+import javax.annotation.ParametersAreNonnullByDefault;
