+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/export/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** Log exporters. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.logs.export;
+
+import javax.annotation.ParametersAreNonnullByDefault;
