+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/package-info.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * The OpenTelemetry SDK implementation of logging.
+ *
+ * @see io.opentelemetry.sdk.logs.LogSinkSdkProvider
+ */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.logs;
+
+import javax.annotation.ParametersAreNonnullByDefault;
