+++ settings.gradle.kts
@@ -45,7 +45,7 @@ include(":exporters:logging-otlp")
 include(":exporters:otlp:all")
 include(":exporters:otlp:common")
 include(":exporters:otlp:metrics")
-include(":exporters:otlp:logging")
+include(":exporters:otlp:logs")
 include(":exporters:otlp:trace")
 include(":exporters:otlp-http:metrics")
 include(":exporters:otlp-http:trace")
@@ -59,7 +59,7 @@ include(":perf-harness")
 include(":proto")
 include(":sdk:all")
 include(":sdk:common")
-include(":sdk:logging")
+include(":sdk:logs")
 include(":sdk:metrics")
 include(":sdk:metrics-testing")
 include(":sdk:testing")
